+++ R/bind.R
@@ -36,6 +36,8 @@ bind_rows_polars <- function(..., .id = NULL) {
 #'  either be a Data/LazyFrame, or a list of Data/LazyFrames. Columns are matched
 #'  by name. All Data/LazyFrames must have the same number of rows and there
 #'  mustn't be duplicated column names.
+#' @param .name_repair Can be `"unique"`, `"universal"`, `"check_unique"`,
+#'  `"minimal"`. See [vctrs::vec_as_names()] for the explanations for each value.
 #'
 #' @export
 #' @examplesIf require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)
@@ -55,8 +57,9 @@ bind_rows_polars <- function(..., .id = NULL) {
 
 bind_cols_polars <- function(
     ...,
-    .name_repair = c("unique", "universal", "check_unique", "minimal")
+    .name_repair = "unique"
   ) {
+  arg_match0(.name_repair, values = c("unique", "universal", "check_unique", "minimal"))
   concat_(..., how = "horizontal", .name_repair = .name_repair)
 }
 
@@ -101,38 +104,47 @@ concat_ <- function(..., how, .id = NULL, .name_repair = NULL) {
   switch(
     how,
     "horizontal" = {
-      # TODO: provide a "name_repair" option
       all_names <- unlist(lapply(dots, names), use.names = FALSE)
       if (anyDuplicated(all_names) > 0) {
         dupes <- get_dupes(all_names)
-        switch(
-          .name_repair,
-          "check_unique" = {
-              msg <- make_dupes_msg(dupes)
-              rlang::abort(
-                c(
-                  "Names must be unique.",
-                  "x" = "These names are duplicated (`variable` (locations)):",
-                  " " = msg
-                ),
-                call = caller_env()
-              )
-          },
-          "minimal" = {
-            rlang::abort(
-              c(
-                "Argument `.name_repair = \"minimal\"` doesn't work on Polars Data/LazyFrames.",
-                "i" = "Either provide unique names or use `.name_repair = \"universal\"`."
-              ),
-              call = caller_env()
-            )
-          },
-          "universal" = {
-            dupes <- names(dupes)
-            vctrs::vec_as_names(dupes, repair = "universal")
-            browser()
+
+        if (.name_repair == "check_unique") {
+          msg <- make_dupes_msg(dupes)
+          rlang::abort(
+            c(
+              "Names must be unique.",
+              "x" = "These names are duplicated (`variable` (locations)):",
+              " " = msg
+            ),
+            call = caller_env()
+          )
+        } else if (.name_repair == "minimal") {
+          rlang::abort(
+            c(
+              "Argument `.name_repair = \"minimal\"` doesn't work on Polars Data/LazyFrames.",
+              "i" = "Either provide unique names or use `.name_repair = \"universal\"`."
+            ),
+            call = caller_env()
+          )
+        } else {
+          # default with quiet = FALSE is *incredibly* slow (more than 60ms)
+          # so I print the message myself
+          new_names <- vctrs::vec_as_names(all_names, repair = .name_repair, quiet = TRUE)
+          mapping <- as.list(all_names)
+          names(mapping) <- new_names
+
+          # make the message myself
+          tmp <- mapping[!names(mapping) %in% all_names]
+          msg <- paste0("`", tmp, "`", " -> ", "`", names(tmp), "`")
+          names(msg) <- rep("*", length(msg))
+          rlang::inform(msg)
+
+          for (i in seq_along(dots)) {
+            n_names <- ncol(dots[[i]])
+            dots[[i]] <- dots[[i]]$rename(mapping[1:n_names])
+            mapping <- mapping[-(1:n_names)]
           }
-        )
+        }
       }
 
       if (inherits(dots[[1]], "DataFrame")) {