+++ tests/tinytest/test_bind_cols_lazy.R
@@ -36,25 +36,6 @@ expect_equal_lazy(
   bind_cols_polars(list(p1, p2))
 )
 
-
-# for now, error if duplicated col names
-
-l2 <- list(
-  polars::pl$LazyFrame(
-    x = sample(letters, 20),
-    y = sample(1:100, 20)
-  ),
-  polars::pl$LazyFrame(
-    y = sample(letters, 20),
-    z = sample(1:100, 20)
-  )
-)
-
-expect_error_lazy(
-  bind_cols_polars(l2),
-  "duplicated names"
-)
-
 l3 <- list(
   data.frame(
     x = sample(letters, 20),
@@ -97,5 +78,55 @@ if (Sys.getenv("TIDYPOLARS_TEST") == "TRUE") {
 }
 
 
+# arg ".name_repair"
+
+l5 <- list(
+  pl$LazyFrame(a = 1, x = 2, y = 3),
+  pl$LazyFrame(z = 1, x = 2, y = 3)
+)
+
+options(rlib_message_verbosity = "quiet")
+
+expect_equal_lazy(
+  names(bind_cols_polars(l5)),
+  c("a", "x...2", "y...3", "z", "x...5", "y...6")
+)
+
+expect_equal_lazy(
+  names(bind_cols_polars(l5, .name_repair = "universal")),
+  c("a", "x...2", "y...3", "z", "x...5", "y...6")
+)
+
+expect_error_lazy(
+  bind_cols_polars(l5, .name_repair = "check_unique"),
+  "Names must be unique"
+)
+
+expect_error_lazy(
+  bind_cols_polars(l5, .name_repair = "minimal"),
+  "Either provide unique names or use"
+)
+
+expect_error_lazy(
+  bind_cols_polars(l5, .name_repair = "blahblah"),
+  "must be one of"
+)
+
+l6 <- list(
+  pl$LazyFrame(x = 1)$rename(list(" " = "x")),
+  pl$LazyFrame(x = 1)$rename(list(" " = "x"))
+)
+
+expect_equal_lazy(
+  names(bind_cols_polars(l6)),
+  c(" ...1", " ...2")
+)
+
+expect_equal_lazy(
+  names(bind_cols_polars(l6, .name_repair = "universal")),
+  c("....1", "....2")
+)
+
+options(rlib_message_verbosity = "default")
 
 Sys.setenv('TIDYPOLARS_TEST' = FALSE)